+++ tests/local_rma.cu
@@ -1,2 +1,1 @@
-#define GHEX_USE_GPU
 #include "local_rma.cpp"
